pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-research-readme"
version = "1.0.0"
description = "Generate professional README files, citations, and automation workflows from simple YAML configuration"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Auto Research README Team", email = "contact@auto-research-readme.com"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Documentation",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Documentation",
    "Topic :: Text Processing :: Markup"
]
keywords = ["readme", "documentation", "research", "automation", "markdown", "citation", "zenodo", "github"]
requires-python = ">=3.9"
dependencies = [
    "PyYAML>=6.0",
    "Jinja2>=3.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.991",
    "isort>=5.0"
]

[project.scripts]
auto-research-readme = "auto_readme.cli:main"

[project.urls]
Homepage = "https://github.com/auto-research-readme/auto-research-readme"
Documentation = "https://github.com/auto-research-readme/auto-research-readme#readme"
Repository = "https://github.com/auto-research-readme/auto-research-readme"
Issues = "https://github.com/auto-research-readme/auto-research-readme/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["auto_readme*"]
exclude = ["tests*"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.pytest_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers"
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests"
] 